Added example from payroll to API Web
