+++ Samples/Program.cs
@@ -1,11 +1,20 @@
 ﻿namespace Samples
 {
+    using WebApiExamples;
+    using MultiIssuerExamples;
+
     class Program
-    {
+    {        
+        /**
+         * La API es una sola pero tiene 2 modalidades (API Web y API Multiemisor):
+         * https://apisandbox.facturama.mx/guias#api-modalidades
+         * Tenemos los ejemplos para ambas, selecciona la que se ajusta a tus necesidades
+         **/
         static void Main(string[] args)
         {
-            EjemplosMultiEmisor.RunExamples();
-            EjemplosWebApi.RunExamples();
+            
+            WebApiExamples.RunExamples();
+            //MultiIssuerExamples.RunExamples();
         }
     }
 }
