+++ Samples/Samples.csproj
@@ -32,9 +32,6 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
-    <Reference Include="Facturama, Version=1.0.8.0, Culture=neutral, processorArchitecture=MSIL">
-      <HintPath>..\packages\Facturama.1.0.8\lib\net452\Facturama.dll</HintPath>
-    </Reference>
     <Reference Include="Newtonsoft.Json, Version=12.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
       <HintPath>..\packages\Newtonsoft.Json.12.0.1\lib\net45\Newtonsoft.Json.dll</HintPath>
     </Reference>
@@ -55,14 +52,24 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="EjemplosMultiEmisor.cs" />
-    <Compile Include="EjemplosWebApi.cs" />
+    <Compile Include="MultiIssuerExamples\MultiIssuerExamples.cs" />
+    <Compile Include="WebApiExamples\CatalogsExample.cs" />
+    <Compile Include="WebApiExamples\InvoiceExample.cs" />
+    <Compile Include="WebApiExamples\PayrollExample.cs" />
+    <Compile Include="WebApiExamples\PaymentComplementExample.cs" />
+    <Compile Include="WebApiExamples\WebApiExamples.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Facturama\Facturama.csproj">
+      <Project>{3d843c81-6a95-4ed6-aa7c-0d5b1424a981}</Project>
+      <Name>Facturama</Name>
+    </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
