+++ Samples/WebApiExamples/WebApiExamples.cs
@@ -1,284 +1,36 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.IO.Ports;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using Facturama;
 using Facturama.Models;
 using Facturama.Models.Request;
-using Facturama.Models.Response.Catalogs.Cfdi;
 
-namespace Samples
+namespace WebApiExamples
 {
-    public class EjemplosWebApi
+    public class WebApiExamples
     {
         public static void RunExamples()
         {
+            // Los datos de usuario y contraseña serán los mismos con los que creaste tu cuenta en Facturama,
+            // Si aún no tienes cuenta en Facturama te recomiendo el crear una en ambiente de sandbox (pruebas)
+            // https://dev.facturama.mx/api/registro
             var facturama = new FacturamaApi("pruebas", "pruebas2011");
-			TestCrudClient(facturama);
-			TestCreateProduct(facturama);
-			TestLogo(facturama);
-			TestSerie(facturama);
-			TestCreateCfdi(facturama);
-			SamplePaymentComplement(facturama);
+                        
+            new CatalogsExample(facturama).Run();               // CRUD  de clientes y productos
+            new InvoiceExample(facturama).Run();                // Creación de factura, descarga de XML y PDF, envío por correo
+            new PaymentComplementExample(facturama).Run();      // Complemento de pago
+            new PayrollExample(facturama).Run();                // Nómina
+
+            // Además puedes editar el logo y agregar series a las sucursales como en los siguientes ejemplos
+            //TestLogo(facturama);
+            //TestSerie(facturama);            
 
             Console.ReadKey();
         }
+        
+        
 
-        private static void TestValidationsClient(FacturamaApi facturama)
-        {
-            try
-            {
-                var cliente = facturama.Clients.Create(new Client
-                {
-                    Address = new Address
-                    {
-                        Country = "MEXICO",
-                        ExteriorNumber = "1230",
-                        InteriorNumber = "B",
-                        Locality = "San Luis",
-                        Municipality = "San Luis Potosí",
-                        Neighborhood = "Lomas 4ta",
-                        State = "San Luis Potosí",
-                        Street = "Cañada de Gomez",
-                        ZipCode = "7822015"
-                    },
-                    CfdiUse = "P0001",
-                    Email = "diego@@facturama.com.mx",
-                    Rfc = "ESO1202108R245",
-                    Name = "Expresion en Software"
-                });
-            }
-            catch (FacturamaException ex)
-            {
-                Console.WriteLine(ex.Message);
-                foreach (var messageDetail in ex.Model.Details)
-                {
-                    Console.WriteLine($"{messageDetail.Key}: {string.Join(",", messageDetail.Value)}");
-                }
-            }
-        }
-
-        private static void TestCrudClient(FacturamaApi facturama)
-        {
-            var clientes = facturama.Clients.List();
-            var clientesBefore = clientes.Count;
-
-            var cliente = facturama.Clients.Create(new Client
-            {
-                Address = new Address
-                {
-                    Country = "MEXICO",
-                    ExteriorNumber = "1230",
-                    InteriorNumber = "B",
-                    Locality = "San Luis",
-                    Municipality = "San Luis Potosí",
-                    Neighborhood = "Lomas 4ta",
-                    State = "San Luis Potosí",
-                    Street = "Cañada de Gomez",
-                    ZipCode = "78220"
-                },
-                CfdiUse = "P01",
-                Email = "diego@facturama.com.mx",
-                Rfc = "ESO1202108R2",
-                Name = "Expresion en Software"
-            });
-
-            cliente = facturama.Clients.Retrieve(cliente.Id);
-            cliente.Rfc = "XAXX010101000";
-            facturama.Clients.Update(cliente, cliente.Id);
-            cliente = facturama.Clients.Retrieve(cliente.Id);
-
-            Console.WriteLine(cliente.Rfc == "XAXX010101000" ? "Cliente Editado" : "Error al editar cliente");
-
-            facturama.Clients.Remove(cliente.Id);
-
-            clientes = facturama.Clients.List();
-            var clientesAfter = clientes.Count;
-
-            Console.WriteLine(clientesAfter == clientesBefore ? "Test Passed!" : "Test Failed!");
-        }
-
-        private static void TestCreateProduct(FacturamaApi facturama)
-        {
-            var unit = facturama.Catalogs.Units("servicio")[0];
-            var prod = facturama.Catalogs.ProductsOrServices("desarrollo")[0];
-            var product = new Product
-            {
-                Unit = "Servicio",
-                UnitCode = unit.Value,
-                IdentificationNumber = "WEB003",
-                Name = "Sitio Web CMS",
-                Description = "Desarrollo e implementación de sitio web empleando un CMS",
-                Price = 6500.0m,
-                CodeProdServ = prod.Value,
-                CuentaPredial = "123",
-                Taxes = new[]
-                {
-                    new Tax
-                    {
-                        Name = "IVA",
-                        Rate = 0.16m,
-                        IsRetention = false,
-                    },
-                    new Tax
-                    {
-                        Name = "ISR",
-                        IsRetention = true,
-                        Total = 0.1m
-                    },
-                    new Tax
-                    {
-                        Name = "IVA",
-                        IsRetention = true,
-                        Total = 0.106667m
-                    }
-                }
-            };
-
-            try
-            {
-                product = facturama.Products.Create(product);
-                Console.WriteLine("Se creo exitosamente un producto con el id: " + product.Id);
-
-                facturama.Products.Remove(product.Id);
-                Console.WriteLine("Se elimino exitosamente un producto con el id: " + product.Id);
-            }
-            catch (FacturamaException ex)
-            {
-                Console.WriteLine(ex.Message);
-                foreach (var messageDetail in ex.Model.Details)
-                {
-                    Console.WriteLine($"{messageDetail.Key}: {string.Join(",", messageDetail.Value)}");
-                }
-            }
-        }
-
-        private static void TestCreateCfdi(FacturamaApi facturama)
-        {
-            var products = facturama.Products.List().Where(p => p.Taxes.Any()).ToList();
-
-            var nameId = facturama.Catalogs.NameIds.ElementAt(1); //Nombre en el pdf: "Factura"
-            var currency = facturama.Catalogs.Currencies.First(m => m.Value == "MXN");
-            var paymentMethod = facturama.Catalogs.PaymentMethods.First(p => p.Name == "Pago en una sola exhibición");
-            var paymentForm = facturama.Catalogs.PaymentForms.First(p => p.Name == "Efectivo");
-            var cliente = facturama.Clients.List().First(c => c.Rfc == "XAXX010101000");
-
-            var branchOffice = facturama.BranchOffices.List().First();
-            var random = new Random();
-            var nitems = random.Next(1, products.Count) % 10 + 1; // Cantidad de items para la factura
-            var decimals = (int)currency.Decimals;
-
-            var cfdi = new Cfdi
-            {
-                NameId = nameId.Value,
-                CfdiType = CfdiType.Ingreso,
-                PaymentForm = paymentForm.Value,
-                PaymentMethod = paymentMethod.Value,
-                Currency = currency.Value,
-                Date = DateTime.Now,
-                ExpeditionPlace = branchOffice.Address.ZipCode,
-                Items = new List<Item>(),
-                Receiver = new Receiver
-                {
-                    CfdiUse = "P01",
-                    Name = cliente.Name,
-                    Rfc = cliente.Rfc
-                },
-            };
-            for (var i = products.Count - nitems; i < products.Count && i > 0; i++)
-            {
-                var product = products[i];
-                var quantity = random.Next(1, 5); //Una cantidad aleatoria
-                var discount = product.Price % (product.Price == 0 ? 1 : random.Next(1, (int)product.Price)); //Un descuento aleatorio
-                var subtotal = Math.Round(product.Price * quantity, decimals);
-
-                var item = new Item
-                {
-                    ProductCode = product.CodeProdServ,
-                    UnitCode = product.UnitCode,
-                    Unit = product.Unit,
-                    Description = string.IsNullOrEmpty(product.Description) ? "Producto de ejemplo" : product.Description,
-                    IdentificationNumber = product.IdentificationNumber,
-                    Quantity = quantity,
-                    Discount = Math.Round(discount, decimals),
-                    UnitPrice = Math.Round(product.Price, decimals),
-                    Subtotal = subtotal,
-                    Taxes = product.Taxes?.Select(
-                        t =>
-                        {
-                            var baseAmount = Math.Round(subtotal - discount, decimals);
-                            return new Tax
-                            {
-                                Name = t.Name,
-                                IsQuota = t.IsQuota,
-                                IsRetention = t.IsRetention,
-
-                                Rate = Math.Round(t.Rate, 6),
-                                Base = Math.Round(subtotal - discount, decimals),
-                                Total = Math.Round(baseAmount * t.Rate, decimals)
-                            };
-                        }).ToList()
-                };
-                var retenciones = item.Taxes?.Where(t => t.IsRetention).Sum(t => t.Total) ?? 0;
-                var traslados = item.Taxes?.Where(t => !t.IsRetention).Sum(t => t.Total) ?? 0;
-                item.Total = item.Subtotal - item.Discount + traslados - retenciones;
-                cfdi.Items.Add(item);
-            }
-
-            try
-            {
-                var cfdiCreated = facturama.Cfdis.Create(cfdi);
-                Console.WriteLine(
-                    $"Se creó exitosamente el cfdi con el folio fiscal: {cfdiCreated.Complement.TaxStamp.Uuid}");
-                facturama.Cfdis.SavePdf($"factura{cfdiCreated.Complement.TaxStamp.Uuid}.pdf", cfdiCreated.Id);
-                facturama.Cfdis.SaveXml($"factura{cfdiCreated.Complement.TaxStamp.Uuid}.xml", cfdiCreated.Id);
-
-                var list = facturama.Cfdis.List("Expresion en Software");
-                Console.WriteLine($"Se encontraron: {list.Length} elementos en la busqueda");
-                list = facturama.Cfdis.List(rfc: "ESO1202108R2"); //Atributo en especifico
-                Console.WriteLine($"Se encontraron: {list.Length} elementos en la busqueda");
-
-                if (facturama.Cfdis.SendByMail(cfdiCreated.Id, "chucho@facturama.mx"))
-                {
-                    Console.WriteLine("Se envió correctamente el CFDI");
-                }
-
-
-                var cancelationStatus = facturama.Cfdis.Cancel(cfdiCreated.Id);
-				if (cancelationStatus.Status == "canceled")
-				{
-					Console.WriteLine($"Se canceló exitosamente el CFDI con el folio fiscal: {cfdiCreated.Complement.TaxStamp.Uuid}");
-				}
-				else if (cancelationStatus.Status == "pending")
-				{
-					Console.WriteLine($"El CFDI está en proceso de cancelacion, require aprobacion por parte del receptor UUID: {cfdiCreated.Complement.TaxStamp.Uuid}");
-				}
-				else if (cancelationStatus.Status == "active")
-				{
-					Console.WriteLine($"El CFDI no pudo ser cancelado, se deben revisar docuementos relacionados on cancelar directo en el SAT UUID: {cfdiCreated.Complement.TaxStamp.Uuid}");
-				}
-				else
-				{
-					Console.WriteLine($"Estado de cancelacin del CFDI desconocido UUID: {cfdiCreated.Complement.TaxStamp.Uuid}");
-				}
-
-			}
-            catch (FacturamaException ex)
-            {
-                Console.WriteLine(ex.Message);
-                foreach (var messageDetail in ex.Model.Details)
-                {
-                    Console.WriteLine($"{messageDetail.Key}: {string.Join(",", messageDetail.Value)}");
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error inesperado: ", ex.Message);
-            }
-        }
-
+        
         private static void TestLogo(FacturamaApi facturama)
         {
             var logo = (new Image
@@ -336,318 +88,8 @@ namespace Samples
                 }
             }
         }
-        /**
-        * Llenado del modelo de CFDI, de una forma general
-        * - Se especifica: la moneda, método de pago, forma de pago, cliente, y lugar de expedición     
-        */
-        private static Facturama.Models.Request.Cfdi CreateModelCfdiGeneral(FacturamaApi facturama)
-        {
-            var products = facturama.Products.List().Where(p => p.Taxes.Any()).ToList();
-
-            var nameId = facturama.Catalogs.NameIds.First(); //Nombre en el pdf: "Factura"
-            var currency = facturama.Catalogs.Currencies.First(m => m.Value == "MXN");
-            var paymentMethod = facturama.Catalogs.PaymentMethods.First(p => p.Value == "PUE");
-            var paymentForm = facturama.Catalogs.PaymentForms.First(p => p.Name == "Efectivo");
-            var cliente = facturama.Clients.List().First();
-
-            var branchOffice = facturama.BranchOffices.List().First();
-
-            var cfdi = new Cfdi
-            {
-                NameId = nameId.Value,
-                CfdiType = CfdiType.Ingreso,
-                PaymentForm = paymentForm.Value,
-                PaymentMethod = paymentMethod.Value,
-                Currency = currency.Value,
-                Date = DateTime.Now,
-                ExpeditionPlace = branchOffice.Address.ZipCode,
-                //Items = new List<Item>(),
-                Receiver = new Receiver
-                {
-                    CfdiUse = cliente.CfdiUse,
-                    Name = cliente.Name,
-                    Rfc = cliente.Rfc
-                },
-            };
-
-            return cfdi;
-        }
-        private static Facturama.Models.Request.Cfdi AddItemsToCfdi(FacturamaApi facturama, Cfdi cfdi)
-        {
-
-            // Lista de todos los productos
-            List<Product> lstProducts = facturama.Products.List();
-            Random random = new Random();
-            var currency = facturama.Catalogs.Currencies.First(m => m.Value == "MXN");
-
-            int nItems = random.Next(lstProducts.Count) % 10 + 1;
-            int decimals = (int)currency.Decimals;
-
-
-            // Lista de Items en el cfdi (los articulos a facturar)
-            List<Item> lstItems = new List<Item>();
-
-            // Creacion del CFDI 
-            for (int i = lstProducts.Count - nItems; i < lstProducts.Count && i > 0; i++)
-            {
-
-                Product product = lstProducts.ElementAt(i);   // Un producto cualquiera
-                int quantity = random.Next(1, 5) + 1;   // una cantidad aleatoria de elementos de este producto
-                Double discount = Decimal.ToDouble(product.Price % (product.Price == 0 ? 1 : random.Next(1, Decimal.ToInt32(product.Price))));
-
-                // Redondeo del precio del producto, de acuerdo a la moneda
-                Double numberOfDecimals = Math.Pow(10, decimals);
-                Double subTotal = Math.Round((Decimal.ToDouble(product.Price) * quantity) * numberOfDecimals) / numberOfDecimals;
-
-
-                // Llenado del item (que va en el cfdi)
-                Item item = new Item
-                {
-                    ProductCode = product.CodeProdServ,
-                    UnitCode = product.UnitCode,
-                    Unit = product.Unit,
-                    Description = $"Item{i}: product.Description",
-                    IdentificationNumber = product.IdentificationNumber,
-                    Quantity = quantity,
-                    Discount = Convert.ToDecimal(Math.Round(discount * numberOfDecimals) / numberOfDecimals),
-                    UnitPrice = Convert.ToDecimal(Math.Round(Decimal.ToDouble(product.Price) * numberOfDecimals) / numberOfDecimals),
-                    Subtotal = Convert.ToDecimal(subTotal)
-
-                };
-
-                // ---- Llenado de los impuestos del item ----                                    
-                item = AddTaxesToItem(item, product, numberOfDecimals);
-                if (item.UnitPrice > 0m)
-                    lstItems.Add(item);
-            }
-            cfdi.Items = lstItems;
-
-            return cfdi;
-        }
-
-        /**
-         * Se agregan los impuestos al Item (uno de los items del cfdi)
-         * Se agregan todos los impuestos del producto, en el caso de que no se tengan impuestos, se debe colocar un valor nulo
-        */
-        private static Item AddTaxesToItem(Item item, Product product, Double numberOfDecimals)
-        {
-            var lstProductTaxes = product.Taxes.Select(i => i).ToList();
-            //List<Tax> lstProductTaxes = product.Taxes.Select(i=>i).ToList(); // impuestos del producto
-            List<Facturama.Models.Request.Tax> lstTaxes = new List<Facturama.Models.Request.Tax>();              // Impuestos del item (del cfdi)
-
-            Double baseAmount = Math.Round((Decimal.ToDouble(item.Subtotal) - Decimal.ToDouble(item.Discount)) * numberOfDecimals) / numberOfDecimals;
-
-            for (int j = 0; j < lstProductTaxes.Count; j++)
-            {
-
-                Facturama.Models.Request.Tax pTax = lstProductTaxes.ElementAt(j);
-
-                Facturama.Models.Request.Tax tax = new Facturama.Models.Request.Tax();
-
-                tax.Name = pTax.Name;
-                tax.IsQuota = pTax.IsQuota;
-                tax.IsRetention = pTax.IsRetention;
-                if (pTax.Rate < 1)
-                {
-                    pTax.Rate = 0;
-                }
-
-                Double rate = Decimal.ToDouble(pTax.Rate);
-                Double rateRounded = (double)Math.Round(rate * 1000000) / 1000000;
-                tax.Rate = Convert.ToDecimal(rateRounded);
-                tax.Base = Convert.ToDecimal(Math.Round(decimal.ToDouble(item.Subtotal) * numberOfDecimals) / numberOfDecimals);
-                tax.Total = Convert.ToDecimal(Math.Round((/*cambie el baseAmount*/decimal.ToDouble(tax.Base * pTax.Rate)) * numberOfDecimals) / numberOfDecimals);
-                if (tax.Base >= 0m)
-                    lstTaxes.Add(tax);
-            }
-
-
-            Double retentionsAmount = 0D;
-            Double transfersAmount = 0D;
-
-            // Asignación de los impuestos, en caso de que no se tengan, el campo va nulo
-            if (lstTaxes.Count > 0)
-            {
-                item.Taxes = lstTaxes;
-
-                retentionsAmount = item.Taxes.Where(o => o.IsRetention).Sum(o => Decimal.ToDouble(o.Total));
-                transfersAmount = item.Taxes.Where(o => !o.IsRetention).Sum(o => Decimal.ToDouble(o.Total));
-            }
-
-            // Calculo del subtotal
-            item.Total = (item.Subtotal - item.Discount + Convert.ToDecimal(transfersAmount) - Convert.ToDecimal(retentionsAmount));
-
-            return item;
-
-        }
-        /**
-        * Modelo "Complemento de pago"
-        * - Se especifica: la moneda, método de pago, forma de pago, cliente, y lugar de expedición     
-        */
-        private static Facturama.Models.Request.Cfdi CreateModelCfdiPaymentComplement(FacturamaApi facturama, Facturama.Models.Response.Cfdi cfdiInicial)
-        {
-
-
-            Cfdi cfdi = new Cfdi();
-
-            // Lista del catálogo de nombres en el PDF
-            var nameForPdf = facturama.Catalogs.NameIds.First(m => m.Value == "14"); // Nombre en el pdf: "Complemento de pago"
-            cfdi.NameId = nameForPdf.Value;
-
-            // Receptor de comprobante (se toma como cliente el mismo a quien se emitió el CFDI Inicial),            
-            String clientRfc = cfdiInicial.Receiver.Rfc;
-            Client client = facturama.Clients.List().Where(p => p.Rfc.Equals(clientRfc)).First();
-
-            Receiver receiver = new Receiver
-            {
-                CfdiUse = "P01",
-                Name = client.Name,
-                Rfc = client.Rfc
-            };
-            cfdi.Receiver = receiver;
-
-            // Lugar de expedición (es necesario por lo menos tener una sucursal)
-            BranchOffice branchOffice = facturama.BranchOffices.List().First();
-            cfdi.ExpeditionPlace = branchOffice.Address.ZipCode;
-
-            // Fecha y hora de expecidión del comprobante
-            //DateTime bindingDate;
-            //DateTime.TryParse(cfdiBinding.Date, null, DateTimeStyles.RoundtripKind, out bindingDate);
-            DateTime cfdiDate = DateTime.Now;
-            cfdi.Date = cfdiDate;
-            cfdi.CfdiType = CfdiType.Pago;
-            // Complemento de pago ---
-            Complement complement = new Complement();
-
-            // Pueden representarse más de un pago en un solo CFDI
-            List<Facturama.Models.Complements.Payment> lstPagos = new List<Facturama.Models.Complements.Payment>();
-            Facturama.Models.Complements.Payment pago = new Facturama.Models.Complements.Payment();
-
-            // Fecha y hora en que se registró el pago en el formato: "yyyy-MM-ddTHH:mm:ss" 
-            // (la fecha del pago debe ser menor que la fecha en que se emite el CFDI)
-            // Para este ejemplo, se considera que  el pago se realizó hace una hora   
-            pago.Date = cfdiDate.AddHours(-1).ToString("yyyy-MM-dd HH:mm:ss");
-
-
-            // Forma de pago (Efectivo, Tarjeta, etc)
-            Facturama.Models.Response.Catalogs.CatalogViewModel paymentForm = facturama.Catalogs.PaymentForms.Where(p => p.Name.Equals("Efectivo")).First();
-            pago.PaymentForm = paymentForm.Value;
-
-            // Selección de la moneda del catálogo
-            // La Moneda, puede ser diferente a la del documento inicial
-            // (En el caso de que sea diferente, se debe colocar el tipo de cambio)
-            List<CurrencyCatalog> lstCurrencies = facturama.Catalogs.Currencies.ToList();
-            CurrencyCatalog currency = lstCurrencies.Where(p => p.Value.Equals("MXN")).First();
-            pago.Currency = currency.Value;
-
-            // Monto del pago
-            // Este monto se puede distribuir entre los documentos relacionados al pago            
-            pago.Amount = 100.00m;
-
-            // Documentos relacionados con el pago
-            // En este ejemplo, los datos se obtiene el cfdiInicial, pero puedes colocar solo los datos
-            // aun sin tener el "Objeto" del cfdi Inicial, ya que los valores son del tipo "String"
-            List<Facturama.Models.Complements.RelatedDocument> lstRelatedDocuments = new List<Facturama.Models.Complements.RelatedDocument>();
-            Facturama.Models.Complements.RelatedDocument relatedDocument = new Facturama.Models.Complements.RelatedDocument
-            {
-                Uuid = cfdiInicial.Complement.TaxStamp.Uuid, // "27568D31-E579-442F-BA77-798CBF30BD7D"
-                Serie = "A",//cfdiInicial.Serie, // "EA"
-                Folio = cfdiInicial.Folio, // 34853
-                Currency = currency.Value,
-                PaymentMethod = "PUE", // En el complemento de pago tiene que ser PUE
-                PartialityNumber = 1,
-                PreviousBalanceAmount = 100.00m,
-                AmountPaid = 100.00m
-            };
-            lstRelatedDocuments.Add(relatedDocument);
-
-            pago.RelatedDocuments = lstRelatedDocuments;
-
-            lstPagos.Add(pago);
-
-            complement.Payments = lstPagos;
-
-            cfdi.Complement = complement;
-
-
-            return cfdi;
-
-        }
-
-        /**
-         * Ejemplo de creación de un CFDI "complemento de pago"
-         * Referencia: https://apisandbox.facturama.mx/guias/api-web/cfdi/complemento-pago
-         * 
-         * En virtud de que el complemento de pago, requiere ser asociado a un CFDI con el campo "PaymentMethod" = "PPD"
-         * En este ejemplo se incluye la creacón de este CFDI, para posteriormente realizar el  "Complemento de pago" = "PUE"     
-        */
-        private static void SamplePaymentComplement(FacturamaApi facturama)
-        {
-
-            try
-            {
-                Console.WriteLine("----- Inicio del ejemplo samplePaymentComplement -----");
-
-                // Cfdi Incial (debe ser "PPD")
-                // -------- Creacion del cfdi en su forma general (sin items / productos) asociados --------
-                Facturama.Models.Request.Cfdi cfdi = CreateModelCfdiGeneral(facturama);
-
-                // -------- Agregar los items que lleva el cfdi ( para este ejemplo, se agregan con datos aleatorios) --------        
-                cfdi = AddItemsToCfdi(facturama, cfdi);
-
-                cfdi.PaymentMethod = "PPD";           // El método de pago del documento inicial debe ser "PPD"
-
-                // Se manda timbrar mediante Facturama
-                Facturama.Models.Response.Cfdi cfdiInicial = facturama.Cfdis.Create(cfdi);
-
-                Console.WriteLine("Se creó exitosamente el cfdi Inicial (PPD) con el folio fiscal: " + cfdiInicial.Complement.TaxStamp.Uuid);
-
-                // Descarga de los archivos del documento inicial
-                String filePath = "factura" + cfdiInicial.Complement.TaxStamp.Uuid;
-                facturama.Cfdis.SavePdf(filePath + ".pdf", cfdiInicial.Id);
-                facturama.Cfdis.SaveXml(filePath + ".xml", cfdiInicial.Id);
-
-
-
-                // Complemento de pago (debe ser "PUE")        
-                // Y no lleva "Items" solo especifica el "Complemento"
-                Facturama.Models.Request.Cfdi paymentComplementModel = CreateModelCfdiPaymentComplement(facturama, cfdiInicial);
-
-
-                // Se manda timbrar el complemento de pago mediante Facturama
-                Facturama.Models.Response.Cfdi paymentComplement = facturama.Cfdis.Create(paymentComplementModel);
-
-                Console.WriteLine("Se creó exitosamente el complemento de pago con el folio fiscal: " + cfdiInicial.Complement.TaxStamp.Uuid);
-
-
-                // Descarga de los archivos del documento inicial
-                String filePathPayment = "factura" + paymentComplement.Complement.TaxStamp.Uuid;
-                facturama.Cfdis.SavePdf(filePath + ".pdf", paymentComplement.Id);
-                facturama.Cfdis.SaveXml(filePath + ".xml", paymentComplement.Id);
-
-
-
-                // Posibilidad de mandar  los cfdis por coreo ( el cfdiInical y complemento de pago)
-                Console.WriteLine(facturama.Cfdis.SendByMail(cfdiInicial.Id, "chucho@facturama.mx"));
-                Console.WriteLine(facturama.Cfdis.SendByMail(paymentComplement.Id, "chucho@facturama.mx"));
-
-
-                Console.WriteLine("----- Fin del ejemplo de samplePaymentComplement -----");
-            }
-            catch (FacturamaException ex)
-            {
-                Console.WriteLine(ex.Message);
-                foreach (var messageDetail in ex.Model.Details)
-                {
-                    Console.WriteLine($"{messageDetail.Key}: {string.Join(",", messageDetail.Value)}");
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"Error inesperado: ", ex.Message);
-            }
-
-        }
+       
+        
     }
 
 }
